test(telemetry): close the coverage gate on the new surfaces
TelemetryEvent::kind() replaces assert!(matches!(..)) in measured tests
(the non-matching arm is a region only a failing assertion reaches);
the runtime observer's tests move to the telemetry/tests.rs layout the
gate excludes; from_config folds the three exporter builds into one
error path (per-exporter early returns were branches only that
exporter's failure reaches) with an unparseable-endpoint test driving
it; RunCompleted's stage-close is unconditional under the open-run
invariant; build_host gets an OTLP-exporter case covering the log-layer
install. cargo xtask coverage: all 11 packages at 100%.
